+++ apps/meteor/client/views/admin/ABAC/AdminABACPage.tsx
@@ -1,7 +1,7 @@
 import { Box, Button, ButtonGroup, Callout } from '@rocket.chat/fuselage';
 import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
-import { GenericModal, ContextualbarDialog, Page, PageContent, PageHeader } from '@rocket.chat/ui-client';
-import { useSetModal, useToastMessageDispatch, useSetting, useEndpoint, useRouteParameter, useRouter } from '@rocket.chat/ui-contexts';
+import { ContextualbarDialog, Page, PageContent, PageHeader } from '@rocket.chat/ui-client';
+import { useSetting, useRouteParameter, useRouter } from '@rocket.chat/ui-contexts';
 import { Trans, useTranslation } from 'react-i18next';
 
 import AttributesContextualBar from './ABACAttributesTab/AttributesContextualBar';
@@ -14,6 +14,7 @@ import RoomsPage from './ABACRoomsTab/RoomsPage';
 import SettingsPage from './ABACSettingTab/SettingsPage';
 import AdminABACTabs from './AdminABACTabs';
 import { useIsABACAvailable } from './hooks/useIsABACAvailable';
+import { useLdapSync } from '../../../hooks/useLdapSync';
 import { useExternalLink } from '../../../hooks/useExternalLink';
 import { links } from '../../../lib/links';
 
@@ -30,45 +31,9 @@ const AdminABACPage = ({ shouldShowWarning }: AdminABACPageProps) => {
 	const learnMore = useExternalLink();
 	const isABACAvailable = useIsABACAvailable();
 	const ldapEnabled = useSetting('LDAP_Enable');
-	const dispatchToastMessage = useToastMessageDispatch();
-	const setModal = useSetModal();
-	const testConnection = useEndpoint('POST', '/v1/ldap.testConnection');
-	const syncNow = useEndpoint('POST', '/v1/ldap.syncNow');
-
-	const handleSyncNowButtonClick = async (): Promise<void> => {
-		try {
-			await testConnection();
-		} catch {
-			dispatchToastMessage({ type: 'error', message: t('Connection_failed') });
-			return;
-		}
-
-		const confirmSync = async (): Promise<void> => {
-			setModal();
-
-			try {
-				const { message } = await syncNow();
-				dispatchToastMessage({ type: 'success', message: t(message as Parameters<typeof t>[0]) });
-			} catch (error) {
-				if (error instanceof Error) {
-					dispatchToastMessage({ type: 'error', message: error });
-				}
-			}
-		};
-
-		setModal(
-			<GenericModal
-				variant='info'
-				confirmText={t('Sync')}
-				cancelText={t('Cancel')}
-				title={t('Execute_Synchronization_Now')}
-				onConfirm={confirmSync}
-				onClose={() => setModal()}
-			>
-				{t('LDAP_Sync_Now_Description')}
-			</GenericModal>,
-		);
-	};
+	const abacEnabled = useSetting('ABAC_Enabled');
+	const handleSyncNow = useLdapSync();
+	const isSyncDisabled = !ldapEnabled || !abacEnabled;
 
 	const handleCloseContextualbar = useEffectEvent((): void => {
 		if (!context) {
@@ -89,7 +54,11 @@ const AdminABACPage = ({ shouldShowWarning }: AdminABACPageProps) => {
 			<Page>
 				<PageHeader title={t('ABAC')}>
 					<ButtonGroup>
-						<Button disabled={!ldapEnabled} onClick={handleSyncNowButtonClick}>
+						<Button
+							disabled={isSyncDisabled}
+							title={isSyncDisabled ? t('Enable_ABAC_and_LDAP_to_sync') : undefined}
+							onClick={handleSyncNow}
+						>
 							{t('LDAP_Sync_Now')}
 						</Button>
 						<Button icon='new-window' secondary onClick={() => learnMore(links.go.abacDocs)}>
